WIP: Base Skip-Chat Component + Wiring into Explorer Ask-Skip Component
