+++ packages/Angular/Explorer/ask-skip/src/lib/dynamic-report/dynamic-drill-down.ts
@@ -1,23 +1,7 @@
 import { Component, Input  } from '@angular/core';
 import { SharedService } from '@memberjunction/ng-shared';
+import { DrillDownInfo } from '@memberjunction/ng-skip-chat';
  
-export class DrillDownInfo {
-  public EntityName!: string;
-  public Filter!: string;
-  public BaseFilter: string = '';
-  public get UserViewGridParams() {
-    const fullFilter = this.BaseFilter?.length > 0 ? `(${this.Filter}) AND (${this.BaseFilter})` : this.Filter;
-    return {
-      EntityName: this.EntityName,
-      ExtraFilter: fullFilter
-    }
-  }
-
-  constructor(entityName: string, filter: string) {
-    this.EntityName = entityName;
-    this.Filter = filter;
-  }
-}
 
 @Component({
   selector: 'mj-dynamic-report-drill-down',
