+++ packages/Angular/Explorer/ask-skip/src/lib/dynamic-report/dynamic-tabbed-report.css
