+++ packages/Angular/Explorer/ask-skip/src/lib/dynamic-report/dynamic-tabbed-report.html
@@ -1,0 +1,106 @@
+@if(SkipData) {
+  <button kendoButton *ngIf="ShowCreateReportButton && !matchingReportID" 
+                      (click)="askCreateReport()" 
+                      [disabled]="isCreatingReport"
+                      class="create-report-button">Create Report</button>
+  <div *ngIf="matchingReportID!==null" class="report-link" (click)="clickMatchingReport()">Report: {{matchingReportName}} (ID: {{matchingReportID}}) Created From This Message</div>
+  @if (ReportEntity && ReportEntity.ID.length > 0) {
+    <div class="report-tab-toolbar">
+      <kendo-button (click)="doRefreshReport()" title="Update this Report with up to date data">
+        <span class="fa-solid fa-arrows-rotate"></span>
+        Refresh
+      </kendo-button>
+    </div>  
+  }
+  <mj-tabstrip class="report-tabstrip" (TabSelected)="onTabSelect($event)" mjFillContainer [bottomMargin]="25" (ResizeContainer)="sharedService.InvokeManualResize()">
+    <mj-tab [TabSelected]="isTabSelected(0)" *ngIf="IsChart">
+      <span class="fa-solid fa-chart-line"></span>
+      <span class="report-tab-title">Chart</span>
+    </mj-tab>
+    <mj-tab-body *ngIf="IsChart">
+      <div class="report-tab-contents" mjFillContainer [bottomMargin]="20" [rightMargin]="20" [style]="{overflowY: 'hidden', overflowX: 'hidden'}">
+        <skip-dynamic-chart mjFillContainer #theChart [AutoResizeChart]="true" [SkipData]="SkipData" [AllowDrillDown]="AllowDrillDown" (DrillDownEvent)="handleChartDrillDown($event)" style="display:block;">
+        </skip-dynamic-chart>  
+      </div>
+    </mj-tab-body>
+
+    <mj-tab [TabSelected]="isTabSelected(1)">
+      <span class="fa-solid fa-table"></span>
+      <span class="report-tab-title">Table</span>
+    </mj-tab>
+    <mj-tab-body>
+      <div class="report-tab-contents"  mjFillContainer [bottomMargin]="20" [rightMargin]="20">
+        <skip-dynamic-grid mjFillContainer #theGrid [SkipData]="SkipData" [AllowDrillDown]="AllowDrillDown" (DrillDownEvent)="handleGridDrillDown($event)">
+        </skip-dynamic-grid>
+      </div>
+    </mj-tab-body>
+
+    <mj-tab *ngIf="AllowDrillDown && DrillDowns.length > 0" [TabSelected]="isTabSelected(2)">
+      <span class="fa-solid fa-table-columns"></span>
+      <span class="report-tab-title">Drill Down</span>
+    </mj-tab>
+    <mj-tab-body *ngIf="AllowDrillDown && DrillDowns.length > 0">
+      <div class="report-tab-contents"  mjFillContainer [bottomMargin]="20" [rightMargin]="20">
+        <mj-dynamic-report-drill-down #drillDownComponent [DrillDowns]="DrillDowns" mjFillContainer style="display:block;"></mj-dynamic-report-drill-down>
+      </div>  
+    </mj-tab-body>
+
+    <mj-tab [TabSelected]="isTabSelected(3)">
+      <span class="fa-solid fa-circle-info"></span>
+      <!-- <img src="assets/Skip - Mark Only - Small.png" class="avatar" style="height:16px;width:16px;" /> -->
+      <span class="report-tab-title">Analysis</span>
+    </mj-tab>
+    <mj-tab-body>
+      <div class="report-tab-contents"  mjFillContainer [bottomMargin]="20" [rightMargin]="15" style="overflow-x: hidden">
+        <markdown [data]="SkipData.analysis"></markdown>
+      </div>
+    </mj-tab-body>
+
+    <mj-tab [TabSelected]="isTabSelected(4)">
+      <span class="fa-solid fa-file"></span>
+      <span class="report-tab-title">Explanation</span>
+    </mj-tab> 
+    <mj-tab-body>
+      <div class="report-tab-contents" mjFillContainer [bottomMargin]="20" [rightMargin]="20">
+        {{SkipData.userExplanation || 'No Explanation Provided'}}
+      </div>
+    </mj-tab-body>
+
+    <mj-tab [TabSelected]="isTabSelected(5)">
+      <span class="fa-solid fa-code"></span>
+      <span class="report-tab-title">Details</span>
+    </mj-tab>
+    <mj-tab-body>
+      <div class="report-tab-contents" mjFillContainer [bottomMargin]="20" [rightMargin]="20">
+        <div *ngIf="ConversationID!==null"><b>Skip Conversation ID:</b><span>{{ConversationID}}</span></div>
+        <div *ngIf="ConversationDetailID!==null"><b>Skip Conversation Detail ID:</b><span>{{ConversationDetailID}}</span></div>
+        <div *ngIf="SkipData?.techExplanation">
+          <span><b>Technical Explanation</b></span>
+          <span><textarea [disabled]="true" [style]="'height: 100px; width: 100%'">{{SkipData.techExplanation}}</textarea></span>
+        </div>
+        <div *ngIf="SkipData?.scriptText" mjFillContainer>
+          <span><b>Script</b></span>
+          <span><textarea mjFillContainer [disabled]="true" [bottomMargin]="20">{{SkipData.scriptText}}</textarea></span>
+        </div>
+      </div>
+    </mj-tab-body>
+  </mj-tabstrip>
+ 
+  <kendo-dialog
+    title="Please confirm"
+    *ngIf="confirmCreateReportDialogOpen"
+    (close)="closeCreateReport('no')"
+    [minWidth]="250"
+    [width]="450"
+  >
+  <p style="margin: 30px; text-align: center;">
+    Would you like to create a new report from this conversation element? If you choose to continue you'll be notified when the report has been created.
+  </p>
+  <kendo-dialog-actions>
+    <button kendoButton (click)="closeCreateReport('no')">No</button>
+    <button kendoButton (click)="closeCreateReport('yes')" themeColor="primary">
+      Yes
+    </button>
+  </kendo-dialog-actions>
+  </kendo-dialog>
+}
