+++ packages/Angular/Explorer/ask-skip/src/public-api.ts
@@ -5,10 +5,7 @@
 export * from './lib/skip-chat/skip-chat.component';
 export * from './lib/skip-chat-with-record/skip-chat-with-record.component';
 export * from './lib/skip-chat-with-record-window/skip-chat-with-record-window.component';
-export * from './lib/dynamic-report/skip-dynamic-report-wrapper';
-export * from './lib/dynamic-report/dynamic-chart';
-export * from './lib/dynamic-report/dynamic-grid';
-export * from './lib/dynamic-report/dynamic-report';
+export * from './lib/dynamic-report/dynamic-tabbed-report';
 export * from './lib/skip-button/skip-button.component';
 export * from './lib/skip-window/skip-window.component';
 export * from './lib/misc/user-view-grid-with-analysis.component'
