+++ packages/Angular/Generic/skip-chat/package.json
@@ -34,8 +34,7 @@
     "angular-plotly.js": "^5.2.2",
     "plotly.js-dist-min": "^2.3.4",
     "@types/plotly.js-dist-min": "^2.3.4",
-    "@angular/cdk": "18.0.2",
-    "ngx-markdown": "18.0.0"
+    "@angular/cdk": "18.0.2"
   },
   "sideEffects": false
 }
